+++ src/collective/aviary/tests/test_controlpanel.py
@@ -11,7 +11,7 @@ from plone.app.testing import logout
 from plone.registry.interfaces import IRegistry
 from zope.component import getUtility
 
-import unittest
+import unittest2 as unittest
 
 
 class ControlPanelTestCase(unittest.TestCase):